Bootstrap 4 support for Pagination
Bootstrap 4's Pagination code is backwards compatible with Bootstrap
3's. It adds two new class names (page-item and page-link) and the
option of aligining the pagination bar using a FlexBox utility class
name. So, these options have been added in a way that works for
Bootstrap 3 and 4.

Pager, however, is simply not supported on Bootstrap 4. It will be
nearly 100% unstyled. It may be removed in a future release as a result.